(feat): separate out PermissionAlert component
